View, Widget, Output: Separated Output buffering from view, widget class
